rename random utils
