+++ pgdrive/utils/__init__.py
@@ -2,6 +2,6 @@ from pgdrive.utils.config import Config, merge_config_with_unknown_keys, merge_c
 from pgdrive.utils.coordinates_shift import panda_heading, panda_position, pgdrive_heading, pgdrive_position
 from pgdrive.utils.cutils import import_cutils
 from pgdrive.utils.math_utils import safe_clip, clip, norm, distance_greater, safe_clip_for_small_array, Vector
-from pgdrive.utils.random import get_np_random, random_string
+from pgdrive.utils.random_utils import get_np_random, random_string
 from pgdrive.utils.utils import is_mac, import_pygame, recursive_equal, setup_logger, merge_dicts, \
     concat_step_infos, is_win
